importFBX general_tab으로 이동
