+++ release/scripts/startup/abler/general_tab/general.py
@@ -452,6 +452,63 @@ class SaveAsOperator(bpy.types.Operator, ExportHelper):
         return {"FINISHED"}
 
 
+class ImportFBXOperator(bpy.types.Operator, AconImportHelper):
+    """Import FBX file according to the current settings"""
+
+    bl_idname = "acon3d.import_fbx"
+    bl_label = "Import FBX"
+    bl_translation_context = "*"
+
+    filter_glob: bpy.props.StringProperty(default="*.fbx", options={"HIDDEN"})
+
+    def execute(self, context):
+        if not self.check_path(accepted=["fbx"]):
+            return {"FINISHED"}
+        try:
+            for obj in bpy.data.objects:
+                obj.select_set(False)
+
+            FILEPATH = self.filepath
+
+            filename = os.path.basename(FILEPATH)
+            col_imported = bpy.data.collections.new(
+                "[FBX] " + filename.replace(".fbx", "")
+            )
+
+            col_layers = bpy.data.collections.get("Layers")
+            if not col_layers:
+                col_layers = bpy.data.collections.new("Layers")
+                context.scene.collection.children.link(col_layers)
+
+            bpy.ops.import_scene.fbx(filepath=FILEPATH)
+            for obj in bpy.context.selected_objects:
+                if obj.name in bpy.context.scene.collection.objects:
+                    bpy.context.scene.collection.objects.unlink(obj)
+                for c in bpy.data.collections:
+                    if obj.name in c.objects:
+                        c.objects.unlink(obj)
+                col_imported.objects.link(obj)
+
+            # put col_imported in l_exclude
+            col_layers.children.link(col_imported)
+            added_l_exclude = context.scene.l_exclude.add()
+            added_l_exclude.name = col_imported.name
+            added_l_exclude.value = True
+
+            # create group
+            bpy.ops.acon3d.create_group()
+            # apply AconToonStyle
+            materials_setup.apply_ACON_toon_style()
+
+        except Exception as e:
+            tracker.import_fbx_fail()
+            raise e
+        else:
+            tracker.import_fbx()
+
+        return {"FINISHED"}
+
+
 class Acon3dGeneralPanel(bpy.types.Panel):
     bl_idname = "ACON3D_PT_general"
     bl_label = "General"
@@ -490,6 +547,10 @@ class Acon3dGeneralPanel(bpy.types.Panel):
         row = layout.row()
         row.operator("acon3d.fly_mode")
 
+        row = layout.row()
+        row.scale_y = 1.0
+        row.operator("acon3d.import_fbx", text="Import FBX")
+
 
 class ApplyToonStyleOperator(bpy.types.Operator):
     """Apply Toon Style"""
@@ -520,6 +581,7 @@ classes = (
     FlyOperator,
     SaveOperator,
     SaveAsOperator,
+    ImportFBXOperator,
 )
